Recup 4 octet in ldi

+++ src/Instructions/ld_ldi.c
@@ -6,7 +6,7 @@
 /*   By: jle-quer <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/25 19:21:53 by jle-quer          #+#    #+#             */
-/*   Updated: 2016/12/12 19:33:29 by jle-quer         ###   ########.fr       */
+/*   Updated: 2016/12/12 20:23:55 by jle-quer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,13 +54,13 @@ void	apply_ldi(t_data *data, t_list *elem, unsigned int *param_types,
 				 param_types[1] == REG_CODE)) && params[2] < 16)
 	{
 		if (param_types[0] == DIR_CODE)
-			value_one = (short)params[0];
+			value_one = (int)params[0];
 		else if (param_types[0] == IND_CODE)
 			value_one = recup_ind(data, (short)params[0], elem->pc);
 		else if (param_types[0] == REG_CODE && params[0] < 16)
 			value_one = elem->reg_number[params[0]];
 		if (param_types[1] == DIR_CODE)
-			value_two = (short)params[1];
+			value_two = (int)params[1];
 		else if (param_types[1] == REG_CODE && params[1] < 16)
 			value_two = elem->reg_number[params[1]];
 		s = (value_one + value_two) % IDX_MOD;
